requirements.txt
asgiref==3.3.1
Django==3.1.7
django-crispy-forms==1.11.1
django-phone-field==1.8.1
pytz==2021.1
sqlparse==0.4.1
